backend/pyproject.toml
[project]
name = "comprende-ya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.0",
    "anthropic>=0.72.0",
    "fastapi[standard]>=0.121.0",
    "openai>=2.7.1",
    "prefect>=3.5.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "yt-dlp>=2025.10.22",
]
